requirements.txt
# Web Framework and API
# Enhanced OSINT Modules - New Dependencies
# Additional AI and ML
# Graph and Visualization
# Security and Forensics
# Additional Intelligence Sources
# Development and Testing
# Database and Search
# Web Dashboard
# Additional dependencies for privacy, logging, and Tor control

# Core Dependencies
requests==2.26.0
beautifulsoup4>=4.11.0
dnspython>=2.2.0
python-whois>=0.7.3
ipwhois>=1.2.0
colorama>=0.4.4
tabulate>=0.9.0
tldextract>=3.4.0
configparser>=5.3.0
urllib3==1.26.7
json5>=0.9.0
rich>=12.6.0
click>=8.1.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
python-dateutil>=2.8.2
pycryptodome>=3.15.0
validators>=0.20.0
netaddr>=0.8.0
phonenumbers>=8.12.0
yara-python>=4.2.0
python-magic>=0.4.27
pytz>=2022.1
pandas==2.1.4
numpy==1.26.4
matplotlib>=3.6.0
seaborn>=0.11.0
networkx>=2.8.0
plotly>=5.10.0
wordcloud>=1.9.0
lxml>=4.9.0
html5lib>=1.1
cryptography>=41.0.0
keyring>=24.0.0
bcrypt>=4.1.0
pytest>=8.0.0
ipaddress>=1.0.8

# Web Framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
starlette>=0.27.0

# Enhanced OSINT Modules
openai>=1.0.0
redis>=5.0.0
elasticsearch>=8.11.0
websockets>=12.0.0
uvloop>=0.17.0
gunicorn>=21.2.0
transformers==4.36.2
huggingface_hub==0.19.4
torch==2.2.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
neo4j>=5.15.0
py2neo>=2021.2.3
lief>=0.13.0
pefile>=2023.2.7
capstone>=5.0.0
unicorn>=2.0.0
waybackpy>=3.0.6
censys==2.1.9
shodan==1.28.0
zoomeye==2.1.2

# Development and Testing
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.3.0
ruff>=0.1.0
psycopg2-binary>=2.9.9
dash>=2.14.0
dash-bootstrap-components>=1.5.0
structlog>=23.1.0
stem>=1.8.0
httpx>=0.24.0
httpx-socks>=0.6.0
respx>=0.20.0
tenacity>=8.2.0

# Enhanced Reporting and PDF Generation
reportlab>=4.0.0
fpdf>=1.7.2
jinja2>=3.1.0
weasyprint>=61.0
markdown>=3.5.0
pillow>=10.0.0

# Advanced ML and Analytics
lightgbm>=4.1.0

# Real-time Processing
celery>=5.3.0
schedule>=1.2.0
apscheduler>=3.10.0
aiosmtplib>=3.0.0
psutil>=5.9.0
